Add files via upload
The gender and age detection model project was an entirely new endeavor for me and one of the most challenging tasks I faced during my Grow Intern internship. Installing the necessary libraries took three days and required watching several tutorials, but persistence paid off, and I eventually succeeded. Understanding the code was particularly interesting because of the new concepts I was learning. This project has been the most fascinating and rewarding one so far. I plan to create more projects using OpenCV, dlib, and DeepFace libraries and hope to move on to Keras and TensorFlow in the future.